キーボード/マウス監視をUIスレッドから逃がすよう修正

+++ VMagicMirror/Assets/Baku/VMagicMirror/Scripts/RawInputObserver/WindowsApiDef.cs
@@ -19,7 +19,7 @@ namespace Baku.VMagicMirror
         public static extern bool SetWindowPos(IntPtr hWnd, int hWndInsertAfter, int X, int Y, int cx, int cy, uint uFlags);
 
         [DllImport("user32.dll")]
-        public static extern IntPtr SetWindowsHookEx(int idHook, HOOKPROC lpfn, IntPtr hMod, IntPtr dwThreadId);
+        public static extern IntPtr SetWindowsHookEx(int idHook, HOOKPROC lpfn, IntPtr hMod, int dwThreadId);
 
         public delegate IntPtr HOOKPROC(int nCode, IntPtr wParam, IntPtr lParam);
